refactor: All fetch requests are moved to a separate file and converted to axios
