+++ frontend/src/pages/Device.vue
@@ -78,7 +78,7 @@
 import { ref, computed, watch, onMounted, onBeforeUnmount } from 'vue'
 import Updater from 'components/Updater.vue'
 import asyncSleep from 'simple-async-sleep'
-import { bytesToSize } from '../util/util'
+import { bytesToSize } from 'util/util'
 import { log } from 'composables/useLog'
 import { rpcErrorHandler } from 'composables/useRpcUtils'
 
@@ -258,15 +258,15 @@ const stopScreenStream = async () => {
   flags.value.screenStream = false
 }
 const onUpdateStage = (stage) => {
-  mainStore.update(stage)
+  mainStore.onUpdateStage(stage)
   if (stage === 'start') {
     flags.value.updateInProgress = true
     stopScreenStream()
     navigator.serial.addEventListener('connect', () => {
-      mainStore.update('end')
+      mainStore.onUpdateStage('end')
     })
   } else if (stage === 'end') {
-    mainStore.update('end')
+    mainStore.onUpdateStage('end')
   }
 }
